requirements.txt
Django==2.2.15
djangorestframework==3.11.1
psycopg2==2.8.5
django-dotenv==1.4.2
requests==2.24.0